requirements.txt
python-telegram-bot==22.7
anthropic
python-dotenv
requests
apscheduler
pytz
youtube-transcript-api
